Added `CustomBackgroundModifier`
Adds a `CustomBackgroundModifier` and a `CustonBackground` enum, plus
some additional helper methods.

The ideia is that we can inject a `CustomBackground` environment value
at a given keypath and the `CustomBackgroundModifier`, which takes a
`KeyPath` in its constructor, will get that value and apply it to the
view using SwiftUI's `background` modifier.